requirements.txt
# django
django>=1.9
djangorestframework==3.5
django-cors-headers==2.0.0

# third
Pillow
requests==2.9.1
APScheduler==3.0.5
python-json-logger==0.1.4
beautifulsoup4==4.4.1
django-model-utils>=2.0.3
jsonfield>=1.0.3
django-crispy-forms
pypinyin
celery==3.1.25
django-celery

# crypt
pycrypto==2.6.1

# date & time
python-dateutil>=2.1
pytz

# database
redis
mysqlclient

# deployment
aiohttp
gunicorn
fabric
